+++ generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/jobs/AbstractJob.java
@@ -15,6 +15,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import lombok.extern.slf4j.Slf4j;
 
+import java.util.Map;
+
 @Getter
 @Slf4j
 @RequiredArgsConstructor
@@ -77,4 +79,16 @@ public abstract class AbstractJob extends AbstractTool {
     }
 
     protected abstract void unsafeExecute() throws GenericException;
+    /**
+     * Method replaces macro expressions like ${FOO} in the input text using dictionary. AstJob's
+     * method doesn't do any replacements as those are to be implemented in its descendants.
+     * For example, Jenkins plugin may override this implementation
+     * with hudson.Util.replaceMacro call
+     * @param value String with macro expressions to be replaced
+     * @param replacements Dictionary with name / value pairs
+     * @return String with macro substitutions complete
+     */
+    public String replaceMacro(@NonNull final String value, final Map<String, String> replacements) {
+        return value;
+    };
 }