+++ ptai-jenkins-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/workmode/subjobs/export/JsonXml.java
@@ -40,6 +40,8 @@ public class JsonXml extends Export {
 
     @Override
     public void apply(@NonNull JenkinsAstJob job) {
+        String fileName = job.replaceMacro(this.fileName);
+        String filter = job.replaceMacro(this.filter);
         Reports.Data data = Reports.Data.builder()
                 .locale(Reports.Locale.valueOf(locale))
                 .format(Reports.Data.Format.valueOf(format))